big-data/server/requirements.txt
flask
flask-cors
cassandra-driver